@@ -1,16 +1,14 @@
 <h1>Counter</h1>
 
 <p>
-    Current count: @currentCount
+    Current count: @_currentCount
 </p>
 
 <button class="btn btn-primary" @onclick="IncrementCount">Click me</button>
 
 @code {
-    int currentCount = 0;
+    int _currentCount;
 
-    void IncrementCount()
-    {
-        currentCount++;
-    }
+    void IncrementCount() 
+        => _currentCount++;
 }
