@@ -10,10 +10,10 @@ namespace Blazored.LocalStorage
         ValueTask ClearAsync(CancellationToken cancellationToken = default);
         bool ContainKey(string key);
         ValueTask<bool> ContainKeyAsync(string key, CancellationToken cancellationToken = default);
-        string GetItem(string key);
-        ValueTask<string> GetItemAsync(string key, CancellationToken cancellationToken = default);
-        string Key(int index);
-        ValueTask<string> KeyAsync(int index, CancellationToken cancellationToken = default);
+        string? GetItem(string key);
+        ValueTask<string?> GetItemAsync(string key, CancellationToken cancellationToken = default);
+        string? Key(int index);
+        ValueTask<string?> KeyAsync(int index, CancellationToken cancellationToken = default);
         IEnumerable<string> Keys();
         ValueTask<IEnumerable<string>> KeysAsync(CancellationToken cancellationToken = default);
         int Length();
